Clippy

+++ src/timeunits.rs
@@ -244,8 +244,8 @@ impl Mul<i64> for Unit {
             Unit::Microsecond => q * (NANOSECONDS_PER_MICROSECOND as i64),
             Unit::Nanosecond => q,
         };
-        if total_ns.abs() < (i64::MAX as i64) {
-            Duration::from_truncated_nanoseconds(total_ns as i64)
+        if total_ns.abs() < i64::MAX {
+            Duration::from_truncated_nanoseconds(total_ns)
         } else {
             Duration::from_total_nanoseconds(total_ns as i128)
         }
